refactor: Renamed function (v1-addon)
