+++ src/migration/v1-addon/steps/create-file-path-maps/addon/map-component-classes.ts
