+++ src/migration/v1-addon/steps/create-file-path-maps/app/map-component-stylesheets.ts
