+++ src/migration/v1-addon/steps/create-file-path-maps/app/map-component-templates.ts
