+++ src/migration/v1-addon/steps/create-file-path-maps/tests/map-components.ts
