+++ src/migration/v1-addon/steps/update-paths-in-app-folder.ts
