+++ tests/helpers/shared-test-setups/ember-v1-addon/typescript.ts
