+++ tests/migration/v1-addon/index/typescript.test.ts
@@ -1,21 +1,21 @@
 import { assertFixture, loadFixture, test } from '@codemod-utils/tests';
 
-import { migrateEmberAddon } from '../../../../src/migration/ember-addon/index.js';
+import { migrateV1Addon } from '../../../../src/migration/v1-addon/index.js';
 import {
   inputProject,
   outputProject,
 } from '../../../fixtures/ember-addon/typescript/index.js';
-import { codemodOptions } from '../../../helpers/shared-test-setups/ember-addon/typescript.js';
+import { codemodOptions } from '../../../helpers/shared-test-setups/ember-v1-addon/typescript.js';
 
-test('migration | ember-addon | index > typescript', function () {
+test('migration | v1-addon | index > typescript', function () {
   loadFixture(inputProject, codemodOptions);
 
-  migrateEmberAddon(codemodOptions);
+  migrateV1Addon(codemodOptions);
 
   assertFixture(outputProject, codemodOptions);
 
   // Check idempotence
-  migrateEmberAddon(codemodOptions);
+  migrateV1Addon(codemodOptions);
 
   assertFixture(outputProject, codemodOptions);
 });
